[android] Fix videoBitrateInMbps
NOTE: video bitrate cannot be determined by a formula based on width and height. Instead, it usually falls into a range of a values (depends on the desire quality and video resolution) but in general "5" is an average value that can work with both 1080p and 720p. Original plugin allows also to leave the field null. We now set it to null